Move login.py to authenticate.py
